+++ NEWS.rst
@@ -62,6 +62,9 @@ warnings anymore when imported.
 A new pairwise sequence aligner is available in Bio.Align, as an alternative
 to the existing pairwise sequence aligner in Bio.pairwise2.
 
+The CodonUsage module now supports the normalized Relative Codon Adaptation
+index, in addition to the Codon Adaptation Index.
+
 Many thanks to the Biopython developers and community for making this release
 possible, especially the following contributors:
 
@@ -71,6 +74,7 @@ possible, especially the following contributors:
 - Chris Rands
 - Connor T. Skennerton
 - Francesco Gastaldello
+- Ivan Erill (first contribution)
 - Michiel de Hoon
 - Pamela Russell (first contribution)
 - Peter Cock